+++ EssentialFeedTests/Feed Cache/FeedStoreCloneSpecs.swift
@@ -1,0 +1,43 @@
+//
+//  FeedStoreCloneSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by Julius on 17/06/2022.
+//
+
+import Foundation
+
+protocol FeedStoreCloneSpecs {
+    func test_retrieve_deliversEmptyOnEmptyCache()
+    func test_retrieve_hasNoSideEffectsOnEmptyCache()
+    func test_retrieve_deliversFoundValuesOnNonEmptyCache()
+    func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
+
+    func test_insert_deliversNoErrorOnEmptyCache()
+    func test_insert_deliversNoErrorOnNonEmptyCache()
+    func test_insert_overridesPreviouslyInsertedCacheValues()
+    
+    func test_delete_deliversNoErrorOnEmptyCache()
+    func test_delete_hasNoSideEffectsOnNonEmptyCache()
+    func test_delete_deliversNoErrorOnNonEmptyCache()
+    func test_delete_emptiesPreviouslyInsertedCache()
+
+    func test_storeSideEffects_runSerially()
+}
+
+protocol FailableRetrieveFeedStoreCloneSpecs: FeedStoreCloneSpecs {
+    func test_retrieve_deliversFailureOnRetrievalError()
+    func test_retrieve_hasNoSideEffectsOnFailure()
+    
+}
+
+protocol FailableInsertFeedStoreCloneSpecs: FeedStoreCloneSpecs {
+    func test_insert_deliversErrorOnInsertionError()
+    func test_insert_hasNoSideEffectsOnInsertionError()
+    
+}
+
+protocol FailableDeleteFeedStoreCloneSpecs: FeedStoreCloneSpecs {
+    func test_delete_deliversErrorOnDeletionError()
+    func test_delete_hasNoSideEffectsOnDeletionError()
+}
